+++ src/Pet4U.Application/UseCases/CreateVolunteer/ICreateVolunteerHandler.cs
@@ -1,5 +1,4 @@
-
-using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Shared;
 
 namespace Pet4U.Application.UseCases.CreateVolunteer;
