+++ src/Pet4U.Application/Validation/CustomValidators.cs
@@ -1,5 +1,5 @@
 using FluentValidation;
-using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Shared;
 
 namespace Pet4U.Application.Validation;
