+++ src/Pet4U.Domain/PetManagement/ValueObjects/PaymentInfoList.cs
@@ -2,7 +2,7 @@ namespace Pet4U.Domain.Volunteers;
 
 public record PaymentInfoList
 {
-  public IReadOnlyCollection<PaymentInfo> Data;
+  public IReadOnlyCollection<PaymentInfo> Data = [];
 
   public PaymentInfoList(IReadOnlyCollection<PaymentInfo> paymentInfos) : base() => Data = paymentInfos;
