+++ src/Pet4U.Domain/PetManagement/ValueObjects/PetPhoto.cs
