+++ src/Pet4U.Domain/PetManagement/ValueObjects/SocialNetwork.cs
