+++ src/Pet4U.Domain/PetManagement/ValueObjects/SocialNetworksList.cs
@@ -1,0 +1,14 @@
+using Pet4U.Domain.Volunteers;
+
+namespace Pet4U.Domain.ValueObjects;
+
+public record SocialNetworksList 
+{
+  public IReadOnlyCollection<SocialNetwork> Data = [];
+
+  public SocialNetworksList(IReadOnlyCollection<SocialNetwork> socialNetworks) => Data = socialNetworks;
+
+
+  // ef core
+  public SocialNetworksList(){}
+}
