+++ src/Pet4U.Domain/PetManagement/ValueObjects/Status.cs
