+++ src/Pet4U.Domain/Shared/Errors.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain.ValueObjects;
+namespace Pet4U.Domain.Shared;
 
 public static class Errors
 {
