+++ src/Pet4U.Domain/Shared/ValueObjects/Description.cs
@@ -1,6 +1,6 @@
-using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Shared;
 
-namespace Pet4U.Domain.ValueObjects;
+namespace Pet4U.Domain.Shared.ValueObjects;
 
 public record Description
 {
