--- src/Pet4U.Domain/Volunteers/SocialNetworksList.cs
@@ -1,14 +1,0 @@
-using Pet4U.Domain.Volunteers;
-
-namespace Pet4U.Domain.ValueObjects;
-
-public record SocialNetworksList 
-{
-  public IReadOnlyCollection<SocialNetwork> Data;
-
-  public SocialNetworksList(IReadOnlyCollection<SocialNetwork> socialNetworks) => Data = socialNetworks;//.ToList();
-
-
-  // ef core
-  public SocialNetworksList(){}
-}
