+++ src/Pet4U.Infrastructure/ApplicationDbContext.cs
@@ -2,8 +2,8 @@ using System.Dynamic;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
-using Pet4U.Domain.Species;
-using Pet4U.Domain.Volunteers;
+using Pet4U.Domain.PetManagement.AgregateRoot;
+using Pet4U.Domain.SpeciesManagement.AgregateRoot;
 
 namespace Pet4U.Infrastructure;
