+++ src/Pet4U.Infrastructure/Configurations/SpeciesConfiguration.cs
@@ -1,8 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Pet4U.Domain.Ids;
-using Pet4U.Domain.Species;
-using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Shared;
+using Pet4U.Domain.Shared.Ids;
+using Pet4U.Domain.SpeciesManagement.AgregateRoot;
 
 namespace Pet4U.Infrastructure 
 {
