+++ src/Pet4U.Infrastructure/Configurations/VolunteerConfiguration.cs
@@ -2,10 +2,9 @@ using System.Data;
 using System.Reflection.Metadata;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Pet4U.Domain;
-using Pet4U.Domain.Ids;
-using Pet4U.Domain.ValueObjects;
-using Pet4U.Domain.Volunteers;
+using Pet4U.Domain.PetManagement.AgregateRoot;
+using Pet4U.Domain.Shared;
+using Pet4U.Domain.Shared.Ids;
 
 namespace Pet4U.Infrastructure 
 {
